refactor(character): split page presentation components
